adjust buildNumericRange interfaces
https://coveord.atlassian.net/browse/KIT-439
