+++ packages/headless/src/features/facets/range-facets/numeric-facet-set/interfaces/request.ts
@@ -1,10 +1,28 @@
-import {
-  RangeRequest,
-  BaseRangeFacetRequest,
-} from '../../generic/interfaces/request';
+import {BaseRangeFacetRequest} from '../../generic/interfaces/request';
 import {CurrentValues, Type} from '../../../facet-api/request';
+import {FacetValueState} from '../../../facet-api/value';
 
-export type NumericRangeRequest = RangeRequest<number>;
+export interface NumericRangeRequest {
+  /**
+   * The start value of the range.
+   */
+  start: number;
+
+  /**
+   * The end value of the range.
+   */
+  end: number;
+
+  /**
+   * Whether to include the `end` value in the range.
+   */
+  endInclusive: boolean;
+
+  /**
+   * The current facet value state.
+   */
+  state: FacetValueState;
+}
 
 export interface NumericFacetRequest
   extends BaseRangeFacetRequest,
